Implement:  Mobile: Describe the plugin settings page #10207
We chose not to proceed with the functionality topic related to GitHub
 mirror because during our analysis of the code in the mobile app, we
did not fully comprehend how integrating with GitHub mirror would be
 beneficial for the project. We tried to fully understand how GitHub
mirror works and how it could be integrated into our project, but we
 concluded that the part of the code related to plugins did not seem
 suitable for using this functionality. As we were unable to clearly
discern the objective or the benefits that GitHub mirror would bring
 in this specific context, we decided to direct our efforts to areas of
the project where we could have a more immediate and
understandable impact.
We focused on the area of plugin categories. First and foremost, it's
important to note that a single plugin can belong to more than one
category. However, upon reviewing the plugins page, we found that
the immediate information available (without clicking on each
plugin's description) seemed insufficient in terms of providing
comprehensive details about the categories each plugin falls into.
Therefore, the first functionality we implemented was on the plugin
search page, displaying one of the categories to which each plugin
belonged. If a plugin belonged to more than one category, we showed
the symbol "+" as there wasn't enough space to list them all. Each
plugin box had to remain concise, and we didn't want to completely
change the developers' project implementation. Additionally, if a
plugin didn't belong to any specific category, we added "Other" to
indicate its classification.
We also added a plugin search filter by category, and on each
specific plugin page, we included all of its categories.

Co-authored-by: Rafael Ribeiro <rafael.m.ribeiro@tecnico.ulisboa.pt>